Funciona todo lo referido con articulos

+++ DeliverEats/Proyecto/Sprints/Sprint 1/Codigo Fuente/DeliverEats/src/main/java/Modules/gestionComercios/controllers/ArticuloController.java
@@ -23,7 +23,7 @@ public class ArticuloController {
      * @param  id  id del articulo a buscar
      * @return      articulo correspondiente a la id
      */
-    public ArticuloModel find(long id) {
+    public ArticuloModel find(int id) {
         T_ARTICULO a = articuloEJB.find(id);
         return convertirArticulo(a);
